Update test docs. Split up auth tests to individual scenario tests.
