--- src/Frontend/test/specs/authentication/auth-full-flow.spec.ts
@@ -1,191 +1,0 @@
-import { vi, expect } from "vitest";
-import { createOidcMock, defaultMockUser } from "../../mocks/oidc-client-mock";
-
-// Mock oidc-client-ts BEFORE any imports that use it
-vi.mock("oidc-client-ts", () => createOidcMock());
-
-import { test, describe } from "../../drivers/vitest/driver";
-import * as precondition from "../../preconditions";
-import { waitFor, screen } from "@testing-library/vue";
-
-describe("FEATURE: Authentication Enabled (Scenario 3)", () => {
-  describe("RULE: Authenticated users should see the dashboard", () => {
-    test("EXAMPLE: Dashboard loads with authenticated user", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify auth token was set by the mocked UserManager
-      const authToken = sessionStorage.getItem("auth_token");
-      expect(authToken).toBe(defaultMockUser.access_token);
-    });
-  });
-});
-
-describe("FEATURE: Token Included in API Requests (Scenario 4)", () => {
-  describe("RULE: Authenticated requests should include Bearer token", () => {
-    test("EXAMPLE: API requests include Authorization header", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      // Track Authorization headers from API requests
-      const capturedHeaders: string[] = [];
-      const serviceControlUrl = window.defaultConfig.service_control_url;
-
-      driver.mockEndpointDynamic(`${serviceControlUrl}endpoints`, "get", (_url, _params, request) => {
-        const authHeader = request.headers.get("Authorization");
-        if (authHeader) {
-          capturedHeaders.push(authHeader);
-        }
-        return Promise.resolve({ body: [] });
-      });
-
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify at least one request included the Bearer token
-      await waitFor(() => {
-        expect(capturedHeaders.length).toBeGreaterThan(0);
-        expect(capturedHeaders[0]).toBe(`Bearer ${defaultMockUser.access_token}`);
-      });
-    });
-  });
-});
-
-describe("FEATURE: Session Persistence (Scenario 6) & Tab Isolation (Scenario 7)", () => {
-  describe("RULE: Session should persist across navigation but be isolated per tab", () => {
-    test("EXAMPLE: Auth token is stored in sessionStorage (tab-specific, not shared)", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify token is in sessionStorage (tab-specific)
-      const sessionToken = sessionStorage.getItem("auth_token");
-      expect(sessionToken).toBe(defaultMockUser.access_token);
-
-      // Verify token is NOT in localStorage (would be shared across tabs)
-      const localToken = localStorage.getItem("auth_token");
-      expect(localToken).toBeNull();
-    });
-
-    test("EXAMPLE: Auth token persists when navigating between pages", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      // Initial navigation - triggers auth
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify initial auth token is set
-      const initialToken = sessionStorage.getItem("auth_token");
-      expect(initialToken).toBe(defaultMockUser.access_token);
-
-      // Navigate to a different route
-      await driver.goTo("/failed-messages/all");
-
-      // Wait for navigation to complete
-      await waitFor(() => {
-        // Verify token persists after navigation
-        const tokenAfterNav = sessionStorage.getItem("auth_token");
-        expect(tokenAfterNav).toBe(defaultMockUser.access_token);
-      });
-
-      // Navigate back to dashboard
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify token still persists
-      const tokenAfterReturn = sessionStorage.getItem("auth_token");
-      expect(tokenAfterReturn).toBe(defaultMockUser.access_token);
-    });
-  });
-});
-
-describe("FEATURE: Logout Flow (Scenario 8)", () => {
-  describe("RULE: Logged-out page should be accessible without authentication", () => {
-    test("EXAMPLE: Logged-out page displays sign-out confirmation", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      // Navigate directly to logged-out page (anonymous route)
-      await driver.goTo("/logged-out");
-
-      await waitFor(() => {
-        // Verify the logged-out page content
-        expect(screen.getByText(/You have been signed out/i)).toBeInTheDocument();
-        expect(screen.getByText(/You have successfully signed out of ServicePulse/i)).toBeInTheDocument();
-        expect(screen.getByRole("button", { name: /Sign in again/i })).toBeInTheDocument();
-      });
-    });
-
-    test("EXAMPLE: Logout clears auth token from session storage", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      // First authenticate
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify token exists
-      expect(sessionStorage.getItem("auth_token")).toBe(defaultMockUser.access_token);
-
-      // Simulate logout by clearing token and navigating to logged-out page
-      sessionStorage.removeItem("auth_token");
-      await driver.goTo("/logged-out");
-
-      await waitFor(() => {
-        expect(screen.getByText(/You have been signed out/i)).toBeInTheDocument();
-      });
-
-      // Verify token is cleared
-      expect(sessionStorage.getItem("auth_token")).toBeNull();
-    });
-  });
-});
-
-describe("FEATURE: Silent Token Renewal (Scenario 9)", () => {
-  describe("RULE: Token renewal should be configured for automatic silent refresh", () => {
-    test("EXAMPLE: UserManager is initialized with silent renewal support", async ({ driver }) => {
-      await driver.setUp(precondition.serviceControlWithMonitoring);
-      await driver.setUp(precondition.hasAuthenticationEnabled());
-
-      await driver.goTo("/dashboard");
-
-      await waitFor(() => {
-        expect(screen.getByText(/Dashboard/i)).toBeInTheDocument();
-      });
-
-      // Verify authentication succeeded (UserManager was initialized and working)
-      // The mocked UserManager includes signinSilent for token renewal
-      expect(sessionStorage.getItem("auth_token")).toBe(defaultMockUser.access_token);
-
-      // Silent renewal is configured via:
-      // - automaticSilentRenew: true in AuthStore.transformToAuthConfig()
-      // - silent_redirect_uri pointing to /silent-renew.html
-      // - Event handlers for addAccessTokenExpiring that call signinSilent()
-      // These are verified by the successful UserManager initialization
-    });
-  });
-});
